% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_tabpfn}
\alias{install_tabpfn}
\title{Install TabPFN and its dependencies}
\usage{
install_tabpfn(envname = "r-tabpfn")
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}
}
\description{
Install TabPFN and its dependencies
}
