% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tabpfn_access_token.R
\name{set_tabpfn_access_token}
\alias{set_tabpfn_access_token}
\title{Set TabPFN Access Token}
\usage{
set_tabpfn_access_token(access_token)
}
\arguments{
\item{access_token}{A string representing the access token.}
}
\description{
This function sets the access token required to authenticate with the TabPFN service.
}
