% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabPFNEstimator.R
\docType{class}
\name{TabPFNBase}
\alias{TabPFNBase}
\title{TabPFN Base Estimator}
\description{
TabPFN Base Estimator

TabPFN Base Estimator
}
\details{
An R6 class to for both TabPFNClassifier and TabPFNRegressor
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{estimator}}{Python TabPFNClassifier or TabPFNRegressor instance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TabPFNBase-new}{\code{TabPFNBase$new()}}
\item \href{#method-TabPFNBase-fit}{\code{TabPFNBase$fit()}}
\item \href{#method-TabPFNBase-predict}{\code{TabPFNBase$predict()}}
\item \href{#method-TabPFNBase-clone}{\code{TabPFNBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNBase-new}{}}}
\subsection{Method \code{new()}}{
Create a new TabPFNClassifier
Fit data, will simply upload the data to the server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNBase$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNBase-fit"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNBase-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNBase$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A data frame or matrix of features.}

\item{\code{y}}{A vector of target values.
Predict the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNBase-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNBase-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNBase$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A data frame or matrix of features.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predictions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
