% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabPFNClassifier.R
\docType{class}
\name{TabPFNClassifier}
\alias{TabPFNClassifier}
\title{TabPFN Classifier}
\description{
An R6 class to interface with the Python TabPFNClassifier.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TabPFNClassifier-new}{\code{TabPFNClassifier$new()}}
\item \href{#method-TabPFNClassifier-fit}{\code{TabPFNClassifier$fit()}}
\item \href{#method-TabPFNClassifier-predict}{\code{TabPFNClassifier$predict()}}
\item \href{#method-TabPFNClassifier-clone}{\code{TabPFNClassifier$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNClassifier-new"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNClassifier-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNClassifier$new(access_token = Sys.getenv("TABPFN_ACCESS_TOKEN"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{access_token}}{Access token for authentication.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNClassifier-fit"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNClassifier-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNClassifier$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A data frame or matrix of features.}

\item{\code{y}}{A vector of target values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNClassifier-predict"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNClassifier-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNClassifier$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{A data frame or matrix of features.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of predictions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNClassifier-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNClassifier-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNClassifier$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
