% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabPFNClassifier.R
\docType{class}
\name{TabPFNRegressor}
\alias{TabPFNRegressor}
\title{TabPFN Regressor}
\description{
TabPFN Regressor

TabPFN Regressor
}
\details{
An R6 class to interface with the Python TabPFNClassifier.
}
\section{Super class}{
\code{\link[tabpfn:TabPFNBase]{tabpfn::TabPFNBase}} -> \code{TabPFNRegressor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TabPFNRegressor-new}{\code{TabPFNRegressor$new()}}
\item \href{#method-TabPFNRegressor-clone}{\code{TabPFNRegressor$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="tabpfn" data-topic="TabPFNBase" data-id="fit"><a href='../../tabpfn/html/TabPFNBase.html#method-TabPFNBase-fit'><code>tabpfn::TabPFNBase$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="tabpfn" data-topic="TabPFNBase" data-id="predict"><a href='../../tabpfn/html/TabPFNBase.html#method-TabPFNBase-predict'><code>tabpfn::TabPFNBase$predict()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNRegressor-new"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNRegressor-new}{}}}
\subsection{Method \code{new()}}{
Create a new TabPFNRegressor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNRegressor$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TabPFNRegressor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TabPFNRegressor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TabPFNRegressor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
